manage gc service in owner
